Write script to download makemeahanzi dictionary
The dictionary data contains other data about a given Hanzi character
that may be of interest to students, such as pinyin and definition.

This data comes from the open source Make Me a Hanzi project:
https://github.com/skishore/makemeahanzi

In particular, this script downloads the following JSONL file:
https://github.com/skishore/makemeahanzi/blob/master/dictionary.txt

And then from there, each line, representing a single character, is
written to the public/characters/ directory to make it possible to
retrieve data only for a given character instead of the entire MBs-worth
of data.