Write script to download makemeahanzi graphics
The graphics data contains data about how to write a given Hanzi
character, including stroke order and radicals.

This data comes from the open source Make Me a Hanzi project:
https://github.com/skishore/makemeahanzi

In particular, this script downloads the following JSONL file:
https://github.com/skishore/makemeahanzi/blob/master/graphics.txt

And then from there, each line, representing a single character, is
written to the public/graphics/ directory to make it possible to
retrieve data only for a given character instead of the entire MBs-worth
of data.